fix: performance issue on archive page
